refactor: move files in src to folders in src
